New employee requests stored on server
Modified functionality so that employee requests data for each employee
is now stored on and retrieved from the server. Newly added requests are
added to the list of requests associated with an employee on the server
and are loaded in with other requests when the employee logs back in.
Number of requests variable is now properly updated as well